Add Spatial Lags of a Variable to a Dataset

Computes spatial lags for a given numeric variable in a dataset using a neighborhood list from the spdep package. It supports inverse distance weighting, exponential, and double power decay weighting methods along with various normalization procedures.

Usage

add_spatial_lags(nblist, y, dat, lags, type = NULL, ...)

Arguments

nblist

An object of class "nb" from the spdep package which represents the neighborhood structure.

y

A character string indicating the name of the numeric variable in dat for which spatial lags will be computed.

dat

A data frame of class "sf" containing the variable specified by y.

lags

A numeric value specifying the number of spatial lags to compute.

type

A character string indicating the type of spatial weights to use. Accepted values are "idw" (inverse distance weighting), "exp" (exponential), "dpd" (double power decay), or NULL (default), which applies standard adjacency-based weighting.

...

Additional arguments passed to spdep::nb2listw() or spdep::nb2listwdist(), such as the style argument specifying the normalization method or zero-policy which indicates whether or not to permit the weights list to be formed with zero-length weights vectors.

Value

A tibble containing the original dataset with additional columns for each computed spatial lag. The spatial lag columns are named "spatlag_<lag>_<y>". The output also includes weight summary attributes named "summ_wgts_spatlag_<lag>".

Details

To obtain a neighborhood list of class "nb," first, a neighborhood algorithm is fit to the geometry of sf dataset. Then, that object is coerced into a neighborhood list. For a workflow, see the Misc section in the Geospatial, Spatial Weights section of my Data Science Notebook.

If type is provided, it matches one of the accepted types ("idw", "exp", "dpd").

If a style argument is passed in ..., it matches one of the valid styles ("W", "B", "C", "S", "U", "minmax", "raw").

The spatial weights summary is extracted from the output of printing the spdep::nb2listw or spdep::nb2listwdist object. It contains characteristics such as the number of regions, number of nonzero links, percentage of nonzero weights, average number of links.

Examples