Create a parameter grid list for dtwclust and dtw distance functions

create_dtw_grids() creates a nested, parameter grid list for dtwclust and dtw distance functions to be used in dtw_dist_gridsearch().

create_dtw_grids(params)

Arguments

params
A named list of parameter name-value pairs. The names should be the names of the distance functions that correspond to the parameter name-value pairs.

Value

Named, nested list where each element is a list of elements that represent each possible configuration for the parameters of that distance function.

Examples

params_ls_lg <- list(dtw_basic = list(window.size = 5:10,
                                      norm = c("L1", "L2"),
                                      step.pattern = list(dtw::symmetric1, dtw::symmetric2)),
                     dtw2 = list(step.pattern = list(dtw::symmetric1, dtw::symmetric2),
                                 window.size = 5:10),
                     dtw_lb = list(window.size = 5:10,
                                   norm = c("L1", "L2"),
                                   dtw.func = "dtw_basic",
                                   step.pattern = list(dtw::symmetric2)),
                     sbd = list(znorm = TRUE, return.shifted = FALSE),
                     gak = list(normalize = TRUE, window.size = 5:10))

dtw_grids_lg <- create_dtw_grids(params_ls_lg)
head(dtw_grids_lg, n = 2)

#> $dtw_basic
#> $dtw_basic[[1]]
#> $dtw_basic[[1]]$window.size
#> [1] 5
#> 
#> $dtw_basic[[1]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[1]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[1]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[2]]
#> $dtw_basic[[2]]$window.size
#> [1] 6
#> 
#> $dtw_basic[[2]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[2]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[2]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[3]]
#> $dtw_basic[[3]]$window.size
#> [1] 7
#> 
#> $dtw_basic[[3]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[3]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[3]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[4]]
#> $dtw_basic[[4]]$window.size
#> [1] 8
#> 
#> $dtw_basic[[4]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[4]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[4]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[5]]
#> $dtw_basic[[5]]$window.size
#> [1] 9
#> 
#> $dtw_basic[[5]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[5]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[5]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[6]]
#> $dtw_basic[[6]]$window.size
#> [1] 10
#> 
#> $dtw_basic[[6]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[6]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[6]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[7]]
#> $dtw_basic[[7]]$window.size
#> [1] 5
#> 
#> $dtw_basic[[7]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[7]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[7]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[8]]
#> $dtw_basic[[8]]$window.size
#> [1] 6
#> 
#> $dtw_basic[[8]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[8]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[8]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[9]]
#> $dtw_basic[[9]]$window.size
#> [1] 7
#> 
#> $dtw_basic[[9]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[9]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[9]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[10]]
#> $dtw_basic[[10]]$window.size
#> [1] 8
#> 
#> $dtw_basic[[10]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[10]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[10]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[11]]
#> $dtw_basic[[11]]$window.size
#> [1] 9
#> 
#> $dtw_basic[[11]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[11]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[11]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[12]]
#> $dtw_basic[[12]]$window.size
#> [1] 10
#> 
#> $dtw_basic[[12]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[12]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw_basic[[12]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw_basic[[13]]
#> $dtw_basic[[13]]$window.size
#> [1] 5
#> 
#> $dtw_basic[[13]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[13]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[13]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[14]]
#> $dtw_basic[[14]]$window.size
#> [1] 6
#> 
#> $dtw_basic[[14]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[14]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[14]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[15]]
#> $dtw_basic[[15]]$window.size
#> [1] 7
#> 
#> $dtw_basic[[15]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[15]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[15]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[16]]
#> $dtw_basic[[16]]$window.size
#> [1] 8
#> 
#> $dtw_basic[[16]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[16]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[16]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[17]]
#> $dtw_basic[[17]]$window.size
#> [1] 9
#> 
#> $dtw_basic[[17]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[17]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[17]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[18]]
#> $dtw_basic[[18]]$window.size
#> [1] 10
#> 
#> $dtw_basic[[18]]$norm
#> [1] "L1"
#> 
#> $dtw_basic[[18]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[18]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[19]]
#> $dtw_basic[[19]]$window.size
#> [1] 5
#> 
#> $dtw_basic[[19]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[19]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[19]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[20]]
#> $dtw_basic[[20]]$window.size
#> [1] 6
#> 
#> $dtw_basic[[20]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[20]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[20]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[21]]
#> $dtw_basic[[21]]$window.size
#> [1] 7
#> 
#> $dtw_basic[[21]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[21]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[21]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[22]]
#> $dtw_basic[[22]]$window.size
#> [1] 8
#> 
#> $dtw_basic[[22]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[22]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[22]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[23]]
#> $dtw_basic[[23]]$window.size
#> [1] 9
#> 
#> $dtw_basic[[23]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[23]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[23]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw_basic[[24]]
#> $dtw_basic[[24]]$window.size
#> [1] 10
#> 
#> $dtw_basic[[24]]$norm
#> [1] "L2"
#> 
#> $dtw_basic[[24]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw_basic[[24]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> 
#> $dtw2
#> $dtw2[[1]]
#> $dtw2[[1]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[1]]$window.size
#> [1] 5
#> 
#> $dtw2[[1]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[2]]
#> $dtw2[[2]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[2]]$window.size
#> [1] 5
#> 
#> $dtw2[[2]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw2[[3]]
#> $dtw2[[3]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[3]]$window.size
#> [1] 6
#> 
#> $dtw2[[3]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[4]]
#> $dtw2[[4]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[4]]$window.size
#> [1] 6
#> 
#> $dtw2[[4]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw2[[5]]
#> $dtw2[[5]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[5]]$window.size
#> [1] 7
#> 
#> $dtw2[[5]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[6]]
#> $dtw2[[6]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[6]]$window.size
#> [1] 7
#> 
#> $dtw2[[6]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw2[[7]]
#> $dtw2[[7]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[7]]$window.size
#> [1] 8
#> 
#> $dtw2[[7]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[8]]
#> $dtw2[[8]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[8]]$window.size
#> [1] 8
#> 
#> $dtw2[[8]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw2[[9]]
#> $dtw2[[9]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[9]]$window.size
#> [1] 9
#> 
#> $dtw2[[9]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[10]]
#> $dtw2[[10]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[10]]$window.size
#> [1] 9
#> 
#> $dtw2[[10]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> $dtw2[[11]]
#> $dtw2[[11]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[11]]$window.size
#> [1] 10
#> 
#> $dtw2[[11]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#> $dtw2[[12]]
#> $dtw2[[12]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] + 2 * d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: N+M
#> 
#> $dtw2[[12]]$window.size
#> [1] 10
#> 
#> $dtw2[[12]]$step_pattern_id
#> [1] "symmetric2"
#> 
#> 
#> 

# Can still be ran with a minimal "grid"
params_ls_sm <- list(dtw2 = list(step.pattern = list(dtw::symmetric1)))

dtw_grids_sm <- create_dtw_grids(params_ls_sm)
head(dtw_grids_sm)

#> $dtw2
#> $dtw2[[1]]
#> $dtw2[[1]]$step.pattern
#> Step pattern recursion:
#> g[i,j] = min(
#>      g[i-1,j-1] +     d[i  ,j  ] ,
#>      g[i  ,j-1] +     d[i  ,j  ] ,
#>      g[i-1,j  ] +     d[i  ,j  ] ,
#>   )
#> 
#>  Normalization hint: NA
#> 
#> $dtw2[[1]]$step_pattern_id
#> [1] "symmetric1"
#> 
#> 
#>